:truck: Move Submission Module to root
